test(conformance): expand expiry + client + connection fixtures with live-Redis parity cases
Grows the conformance harness fixtures and the smoke-test case
lists for core expiry, client, connection, errors, generic, object,
server, and strings domains so the corners that the new CLIENT
TRACKING / CACHING / PAUSE and EXPIRETIME parity work uncovered are
locked in against live Redis via the live-diff comparator.

Fixture additions (crates/fr-conformance/fixtures/core_*.json) cover
things like:

- Missing/no-expiry observability contracts for PTTL/TTL/PERSIST/
  EXPIRETIME/PEXPIRETIME: -1 on no-expiry, -2 on missing key, 0/1 on
  EXPIRE/PERSIST no-op vs applied.
- Relative expiry option matrices for PEXPIRE: NX/XX/GT/LT interactions
  so the new bespoke error strings and non-application branches are
  all exercised.
- CLIENT TRACKING / CACHING / TRACKINGINFO live-diff cases that exercise
  the new switch-mode errors, caching-yes/caching-no flag emission, and
  the updated CLIENT PAUSE / UNBLOCK error phrasing.

crates/fr-conformance/src/lib.rs:

- Updates the expire_variant test to expect EXPIRETIME = 6 (not 5)
  once the absolute-ms -> seconds conversion uses (abs_ms + 500) / 1000
  rounding. The comment is rewritten so the next reader sees Redis's
  rounding rule rather than the old "truncates" assumption.

crates/fr-conformance/tests/smoke.rs:

- Adds CORE_EXPIRY_LIVE_STABLE_CASES and a matching BTreeMap-backed
  set-up so the expiry fixtures participate in the live-diff harness
  alongside the previously covered client / connection / errors /
  generic / object / server / strings surfaces. The case list
  mirrors the fixture names one-to-one; if a fixture starts drifting
  from live Redis, smoke.rs now points at exactly which case failed.

+++ crates/fr-conformance/src/lib.rs
@@ -5215,8 +5215,10 @@ mod tests {
         assert_eq!(expire_variant, pexpire_variant);
         assert_eq!(expire_variant.0, RespFrame::Integer(5));
         assert_eq!(expire_variant.1, RespFrame::Integer(5000));
-        // expires_at_ms = 840 + 5000 = 5840; EXPIRETIME truncates: 5840/1000 = 5
-        assert_eq!(expire_variant.2, RespFrame::Integer(5));
+        // expires_at_ms = 840 + 5000 = 5840; Redis converts absolute
+        // millisecond deadlines with (abs_ms + 500) / 1000, so 5840ms
+        // reports as second 6.
+        assert_eq!(expire_variant.2, RespFrame::Integer(6));
         assert_eq!(expire_variant.3, RespFrame::Integer(5840));
 
         let event = EvidenceEvent {
